requirements.txt
annotated-types==0.7.0
asttokens==2.4.1
attrs==25.1.0
bitarray==3.0.0
black==25.1.0
cached-property==2.0.1
cbor2==5.6.5
certifi==2025.1.31
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
coverage==7.6.10
cytoolz==1.0.1
eth-account==0.13.4
eth-bloom==3.1.0
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.1.0
eth-stdlib==0.2.8
eth-typing==5.1.0
eth-utils==5.2.0
eth_abi==5.2.0
flake8==7.1.1
hexbytes==1.3.0
hypothesis==6.125.1
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
isort==6.0.0
lark==1.2.2
lru-dict==1.3.0
mamushi==0.0.4
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==0.4.3
packaging==23.2
parsimonious==0.10.0
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
py-ecc==7.0.1
py-evm==0.10.1b2
pycodestyle==2.12.1
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.1
pytest==8.3.4
pytest-cov==6.0.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rlp==4.0.1
six==1.17.0
sortedcontainers==2.4.0
titanoboa==0.2.5
toolz==1.0.0
trie==3.1.0
typing_extensions==4.12.2
urllib3==2.3.0
vvm==0.3.2
vyper==0.4.0
wheel==0.45.1
zipp==3.21.0
